#9744: add pybind for ttnn pad
