+++ tt_eager/tt_dnn/op_library/composite/composite_ops.hpp
@@ -15,7 +15,7 @@
 
 
 #include "ttnn/operations/eltwise/binary/device/binary_op.hpp"
-#include "ttnn/operations/data_movement/pad.hpp"
+#include "ttnn/operations/data_movement/pad/pad.hpp"
 
 namespace tt {
