+++ ttnn/cpp/pybind11/operations/__init__.hpp
@@ -11,7 +11,6 @@
 #include "pybind11/operations/conv2d.hpp"
 #include "pybind11/operations/core.hpp"
 #include "pybind11/operations/creation.hpp"
-#include "pybind11/operations/data_movement.hpp"
 #include "pybind11/operations/embedding.hpp"
 #include "pybind11/operations/kv_cache.hpp"
 #include "pybind11/operations/matmul.hpp"
@@ -26,6 +25,7 @@
 #include "ttnn/operations/eltwise/unary/unary_pybind.hpp"
 #include "ttnn/operations/reduction/reduction_pybind.hpp"
 #include "ttnn/operations/eltwise/binary_backward/binary_backward_pybind.hpp"
+#include "ttnn/operations/data_movement/data_movement_pybind.hpp"
 
 
 namespace py = pybind11;
